Gutenboarding: Move verticals reducer into separate store
