+++ client/landing/gutenboarding/stores/onboard/reducer.ts
@@ -7,7 +7,7 @@ import { combineReducers } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import { ActionType, Vertical, SiteVertical } from './types';
+import { ActionType, SiteVertical } from './types';
 import { DomainSuggestion } from '../domain-suggestions/types';
 import * as Actions from './actions';
 
@@ -31,16 +31,6 @@ const siteTitle: Reducer< string, ReturnType< typeof Actions[ 'setSiteTitle' ] >
 	return state;
 };
 
-const verticals: Reducer< Vertical[], ReturnType< typeof Actions[ 'receiveVerticals' ] > > = (
-	state = [],
-	action
-) => {
-	if ( action.type === ActionType.RECEIVE_VERTICALS ) {
-		return action.verticals;
-	}
-	return state;
-};
-
 const siteVertical: Reducer<
 	SiteVertical | undefined,
 	ReturnType< typeof Actions[ 'setSiteVertical' ] >
@@ -54,7 +44,7 @@ const siteVertical: Reducer<
 	return state;
 };
 
-const reducer = combineReducers( { domain, siteTitle, verticals, siteVertical } );
+const reducer = combineReducers( { domain, siteTitle, siteVertical } );
 
 export type State = ReturnType< typeof reducer >;
