+++ client/landing/gutenboarding/stores/onboard/types.ts
@@ -1,5 +1,4 @@
 enum ActionType {
-	RECEIVE_VERTICALS = 'RECEIVE_VERTICALS',
 	RESET_SITE_TYPE = 'RESET_SITE_TYPE',
 	SET_DOMAIN = 'SET_DOMAIN',
 	SET_SITE_TITLE = 'SET_SITE_TITLE',
@@ -12,19 +11,4 @@ export { ActionType };
 export interface SiteVertical {
 	label: string;
 	id: string;
-}
-
-/**
- * Representation of well-known verticals
- */
-export interface Vertical {
-	is_user_input_vertical: false;
-
-	vertical_id: string;
-	vertical_slug: string;
-	vertical_name: string;
-	parent: string;
-	preview: string;
-	preview_styles_url: string;
-	synonyms: string[];
 }
