+++ client/landing/gutenboarding/stores/verticals/actions.ts
@@ -1,0 +1,9 @@
+/**
+ * Internal dependencies
+ */
+import { ActionType, Vertical } from './types';
+
+export const receiveVerticals = ( verticals: Vertical[] ) => ( {
+	type: ActionType.RECEIVE_VERTICALS as const,
+	verticals,
+} );
