+++ client/landing/gutenboarding/stores/verticals/constants.ts
@@ -1,0 +1,1 @@
+export const STORE_KEY = 'automattic/verticals';
