+++ client/landing/gutenboarding/stores/verticals/index.ts
@@ -2,7 +2,7 @@
  * External dependencies
  */
 import { controls } from '@wordpress/data-controls';
-import { plugins, registerStore, use } from '@wordpress/data';
+import { registerStore } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -11,18 +11,17 @@ import { STORE_KEY } from './constants';
 import reducer, { State } from './reducer';
 import * as actions from './actions';
 import * as selectors from './selectors';
+import * as resolvers from './resolvers';
 import { SelectFromMap, DispatchFromMap } from '../mapped-types';
 
 export { STORE_KEY };
-
-use( plugins.persistence, {} );
 
 registerStore< State >( STORE_KEY, {
 	actions,
 	controls,
 	reducer,
+	resolvers,
 	selectors,
-	persist: [ 'domain', 'siteTitle', 'siteVertical' ],
 } );
 
 declare module '@wordpress/data' {
