+++ client/landing/gutenboarding/stores/verticals/resolvers.ts
