+++ client/landing/gutenboarding/stores/verticals/selectors.ts
@@ -4,3 +4,4 @@
 import { State } from './reducer';
 
 export const getState = ( state: State ) => state;
+export const getVerticals = ( state: State ) => state.verticals;
