+++ client/landing/gutenboarding/stores/verticals/types.ts
@@ -1,0 +1,19 @@
+enum ActionType {
+	RECEIVE_VERTICALS = 'RECEIVE_VERTICALS',
+}
+export { ActionType };
+
+/**
+ * Representation of well-known verticals
+ */
+export interface Vertical {
+	is_user_input_vertical: false;
+
+	vertical_id: string;
+	vertical_slug: string;
+	vertical_name: string;
+	parent: string;
+	preview: string;
+	preview_styles_url: string;
+	synonyms: string[];
+}
